Update Robot.java
This is an example for retrieving values from the NavX or a Spark Max encoder and subsequently sending them to a dashboard. The values should be visible on both SmartDashboard and Shuffleboard (depending on which one you select from the Driver Station). The code for a "sendable chooser" was pre-loaded into the template, but could be useful for selecting different Autonomous modes.